backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-RESTX==1.3.0
Flask-SocketIO==5.3.5

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
Flask-Caching==2.1.0

# Background Tasks
celery==5.3.4

# WebSocket
python-socketio==5.10.0
eventlet==0.33.3

# Data Processing
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33

# API Clients
requests==2.31.0
aiohttp==3.9.1

# Security
bcrypt==4.1.2
python-dotenv==1.0.0
werkzeug==3.0.1

# Data Validation
marshmallow==3.20.1
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing (Development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Code Quality (Development)
black==23.12.1
flake8==6.1.0
mypy==1.7.1
